requirements.txt
django
openai-agents
python-dotenv 
openai
django-environ
django-cors-headers
django-after-response
certifi
aiohttp 
aiofiles
markdown
beautifulsoup4
<<<<<<< HEAD
python-docx
=======
python-docx
uvicorn
>>>>>>> 298b180593eec7888be759fd446b324f2cd91908
